chore: Update vscode settings and tasks for pytest
This commit updates the vscode settings and tasks to enable pytest for testing. The "files.exclude" setting is removed to include all files in the workspace. The "python.testing.unittestEnabled" setting is removed as it is no longer needed. The "python.testing.pytestEnabled" setting is set to true to enable pytest. Additionally, a new task "Run pytest" is added to the tasks.json file, which runs the pytest command and presents the test results in a new panel. This update improves the testing workflow and makes it easier to run and view pytest results in vscode.